update to make text colors different
Each time user visits the chat, he/she receives unique random color of message stickers and labels. Not often but it happens that user takes dark background, as a result it's hard to read the message text of this user. So to make reading of the messages more convenient I made mechanism to choose text color white or black on base of background brightness.